chore: remove the LICENSE file
After internal discussions, it was decided that no explicit LICENSE file
is needed in this repository. The one that is removed in this commit was
not really a license file anyway.

The language in the `CONTRIBUTING.md` file was updated to make a
distinction between internal and external contributors. Due to the
possible legal gray area around accepting external code contributions,
the language was updated to make it more clear that community
contributions should be limited to submissions/discussions in the issue
tracker.