docs: add a FAQ page
This change adds a Frequently Asked Questions (FAQ) page and populates
it with some initial content. The page has been added under the
"Knowledge Base" category and the first question comes from the current
"Processing" page hosted at https://docs.phylum.io/docs/processing

That processing page is not currently tracked in source control. It only
lives on the readme.com site. The other page like that is the "Phylum
Package Score" at https://docs.phylum.io/docs/phylum-package-score

That page will not be ported since the decision was to abandon it given
the content will not be relevant going forward.

Closes #30