docs: add initial content
This is a WIP commit to get the content checked into the repo in case
someone else wants to work on it. The relevant pages from the
`brad/guides` branch were added to this branch and modified to get a
start on the "Defend Your Workstation" (DYW) page. The top-level DYW
page is still needed, but it can reference the
`protected_package_install.md` doc added here, which in turn references
the newly added `create_project.md` doc.