docs: add the "Defend Your Workstation" page
This page serves as a text-based landing page. It may be adapted to a
visual (e.g., clickable image tiles) page after the switch to
Docusaurus. The content is action oriented, with compelling user stories
to help sell the value offered by the integrations. Each of the current
options are listed, in order from higher to lower levels of security.
This format also intends to show the ability to shift software supply
chain security further and further left.