docs: update GitHub app doc to match UI changes (#58)
This change updates the screenshots that were out of date with the
latest UI refresh. The text was also updated to be current. This PR
comes on the heels of the update to the GitHub App [marketplace
listing](https://github.com/marketplace/phylum-io), which now directs
more prominently to this documentation page.