WIP: update submodule to point to a branch
While testing the changes on the `cli` repository, it is better to have
the git submodule for it point to a specific branch instead of the
default one (`main`). The branch used here is named `new_docs_site` and
will be temporary. Once the changes in both repositories are accepted,
the submodule will be reverted back to point at the default branch.