fix: show correct last update time/author (#90)
This PR addresses the issue of the "Last updated on" time and author
being the same for every page and set to the most recent commit
time/author. It modifies the git fetch depth in the `docs` workflow used
for publishing the site, as detailed in this discussion:
https://github.com/facebook/docusaurus/issues/2798#issuecomment-636602951.

This PR also takes the opportunity to update the Discord Phylum
Community link so that it shows "Louis (lwlang)" as the inviter instead
of "Bill@Phylum (bill_phylum)".

Finally, a few grammar updates were made and a few typos fixed.