ci: update submodules in submodule update workflow (#98)
The change in #97 was not the solution as a manual trigger to the
updated workflow resulted in the same error. Further investigation
revealed that the submodule was not updated (did not have the latest
commits and tags) before the workflow attempted to checkout the newly
created tag.

This change accounts for that update step. It also ensures that each
submodule command in the workflow and in documentation is acting on the
explicitly named submodule so that other submodules (which don't exist
yet but may soon) are not accidentally changed/updated.

Finally, the `actions/checkout` option to `fetch-tags` was removed since
it was determined to act on the `documentation` repo itself and not any
of the submodules. The `documentation` repo does not have any tags and,
even if it did, they aren't needed for the submodule update workflow.

---------

Co-authored-by: Kyle Willmon <kyle@phylum.io>