docs: use submodule for `phylum-ci` integrations
This change removes the integration documents that are in provided by
the `phylum-ci` repository in favor of pulling those same documents in
as a git submodule instead. That allows for the code and the docs to
exist together, be updated together, and ultimately syncronized in this
repository only when `phylum-ci` has a new release. Changes include:

* Remove integration documents
* Add a new `ext/phylum-ci` submodule
* Create symlink from submodule: `docs/integrations` to `docs/phylum-ci`
* Update existing links to point to the new path
* Update the sidebar configuration to ensure pages show in same order
  * (This also fixed the "Integrations Overview" page showing twice)
* Ensure the "edit this page" link works for the `phylum-ci` pages
* Add client side redirect entries for each changed location