update policy documentation for v1 (#152)
This PR updates the policy documentation to reflect the new `policy.v1`
format. The previous format still works, but the new format is
preferred.

There's currently an issue with the policy SDK zip where the root
element is missing. I'm pretty sure it was unintentional and there isn't
a trick to making the opa CLI load data with a prefix. It should be like
this one:


[constants.json](https://github.com/user-attachments/files/15537760/constants.json)


## Checklist

- [ ] Does this PR have an associated issue (i.e., `closes #<issueNum>`
in description above)?
- [ ] Have you ensured that you have met the expected acceptance
criteria?