docs: add a new FAQ entry
This new entry is the result of the question emerging from several users
of the integrations. They didn't understand that the Phylum UI shows the
status of the entire project while integrations may only show the status
of the new/modified dependency files...the diffs. This FAQ page will be
linked from the other sources where the question may come up.