Update Aviary API docs to reference Swagger
This patch makes use of the new shared Swagger instance to avoid
manually documenting artifact repository endpoints.

It also adds a new authentication section which can be referenced by
Swagger in the future, to help with generating a valid basic auth token.

The webhook checksum section was expanded slightly to make it clear that
the purpose of the checksum is validating the request, rather than just
documenting that the checksum is provided.