Bump `docusaurus` to v3.6.3
This change bumps the version of `docusaurus` from v3.5.2 to v3.6.3.

The new "Docusaurus Faster" feature added in v3.6.0 was applied with a
feature flag in the config. It makes use of SWC js loader, which allows
for removing the Babel config file at `site/babel.config.js`. It also
uses `rspack` instead of `webpack`, which required a few small changes
to the config (mostly naming and comment reminders). Overall, using the
new features allows the site to build in one third the time as before,
at least when measured from local builds.

The changes were tested locally by successfully building and serving the
site. Changelogs for the versions included in this upgrade:

* [v3.6.0](https://github.com/facebook/docusaurus/releases/tag/v3.6.0)
* [v3.6.1](https://github.com/facebook/docusaurus/releases/tag/v3.6.1)
* [v3.6.2](https://github.com/facebook/docusaurus/releases/tag/v3.6.2)
* [v3.6.3](https://github.com/facebook/docusaurus/releases/tag/v3.6.3)