copy artifactory page for nexus (#274)
This is basically a copy of the Artifactory documentation adapted for
Nexus. These instructions are applicable for all the supported
ecosystems, but so far I have only tested restoring packages from PyPI
and Cargo.

I think it's a good idea to recommend reducing the cached time, but I
don't have advice on how the user should determine a more appropriate
value.

## Checklist

- [ ] Does this PR have an associated issue (i.e., `closes #<issueNum>`
in description above)?
- [ ] Have you ensured that you have met the expected acceptance
criteria?

---------

Co-authored-by: Charles Coggins <maxrake@users.noreply.github.com>