add IE1001 and CM1002 (#265)
This PR adds the new IE1001 and CM1002 issue tags to the issue tag
table.

I didn't add pages to the analytics section. Should I? Most of the other
tags have one.

IE1001 says it's for RustSec, but I think in the future it could be
expanded to other ecosystems if there are other ecosystems that use a
similar system for marking packages as unmaintained. It's similar to
CM0046 Deprecated Package, except that unmaintained doesn't necessarily
mean that there's anything really wrong with the package or that a
reasonable alternative exists, and the data is sourced in a different
way.

## Checklist

- [ ] Does this PR have an associated issue (i.e., `closes #<issueNum>`
in description above)?
- [ ] Have you ensured that you have met the expected acceptance
criteria?

---------

Co-authored-by: Charles Coggins <maxrake@users.noreply.github.com>