Bump `docusaurus` to v3.7.0 (#289)
This change bumps the version of `docusaurus` from v3.6.3 to v3.7.0.
There are no new features or bug fixes that need to be accounted for.
The changes were tested locally by successfully building and serving the
site. Changelogs for the versions included in this upgrade:

* [v3.7.0](https://github.com/facebook/docusaurus/releases/tag/v3.7.0)

The version of `react` was updated at this time to v19.x since this
release of `docusaurus` suggests it, prior to the next release when it
will be mandatory.

Additionally, this change updates the weekly dependency update workflow,
to bring it up to date with the rest of the org and also allow for an
earlier update time so that it will be more likely that all packages are
processed by the time someone reviews the changes.