Update Poetry package firewall command and output (#295)
This change updates the command to use to add the `phylum` source repo
so that it makes use of the `--priority primary` option, which will be
required in a future release of Poetry and makes it more explicit now.
The output when **all** versions of a package are blocked is also
updated to match the current output text when this happens.