Bump `docusaurus` to v3.8.1
This change bumps the version of `docusaurus` from v3.7.0 to v3.8.1.
There are no new features or bug fixes that need to be accounted for.
The changes were tested locally by successfully building and serving the
site. Changelogs for the versions included in this upgrade:

* [v3.8.0](https://github.com/facebook/docusaurus/releases/tag/v3.8.0)
* [v3.8.1](https://github.com/facebook/docusaurus/releases/tag/v3.8.1)

The next major version of Docusaurus (v4) is coming and there are
[Future Flags](https://docusaurus.io/blog/releases/3.8#future-flags)
that can be opted into to prepare for that release. All of those flags
were accepted.