Remove references to the free Community version (#310)
This change was initiated due to a potential customer reading the
documentation and following the link to register for the free version.
When it didn't work for them, they used the contact info published on
the site to ask about the issue. I informed them of the new model of
operations and told them we would update the documentation to match.
This is that update.