Bump `docusaurus` to v3.9.1
This change bumps the version of `docusaurus` from v3.8.1 to v3.9.1.
There are no new features or bug fixes that need to be accounted for.
The changes were tested locally by successfully building and serving the
site. Changelogs for the versions included in this upgrade:

* [v3.9.0](https://github.com/facebook/docusaurus/releases/tag/v3.9.0)
* [v3.9.1](https://github.com/facebook/docusaurus/releases/tag/v3.9.1)

The next major version of Docusaurus (v4) is coming and the config
setting `siteConfig.onBrokenMarkdownLinks` will be removed in favor of
`siteConfig.markdown.hooks.onBrokenMarkdownLinks`. The change was made
now since it was deprecated in this upgrade. Additionally, a new setting
for `siteConfig.markdown.hooks.onBrokenMarkdownImages` was introduced
and adopted here.